# Conflicts:
#	longjian-measure-app/src/main/java/com/longfor/longjian/measure/app/appService/appService/IAPPMeasureService.java

+++ longjian-measure-app/src/main/java/com/longfor/longjian/measure/app/vo/appMeasureSyncVo/RepairerListVo.java
@@ -31,9 +31,9 @@ public class RepairerListVo {
     /**
      * 更新时间
      */
-    private Long update_at;
+    private Integer update_at;
     /**
      * 删除时间(0表示未删除)
      */
-    private Long delete_at;
+    private Integer delete_at;
 }
