+++ longjian-measure-app/src/main/java/com/longfor/longjian/measure/app/vo/appMeasureSyncVo/ResultListVo.java
@@ -57,15 +57,15 @@ public class ResultListVo {
     /**
      * 得分
      */
-    private Float score;
+    private Double score;
     /**
      * 更新时间
      */
-    private Long update_at;
+    private Integer update_at;
     /**
      * 删除时间(0表示未删除)
      */
-    private Long delete_at;
+    private Integer delete_at;
     /**
      * 描区uuid
      */
